refactor: pages 폴더를 page-components로 이름 변경
문제:
- Next.js가 src/pages를 Pages Router로 인식
- App Router 사용 중인데 충돌 발생

해결:
- src/pages → src/page-components 폴더명 변경
- 모든 import 경로 업데이트 (@/pages → @/page-components)
- tsconfig.json paths 업데이트
- 22개 app 페이지 파일의 import 경로 수정

결과:
- 빌드 성공
- Vercel 배포 가능
