+++ src/app/bmc/[id]/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import BmcDetailPage from '@/pages/Bmc/BmcDetailPage'
+import BmcDetailPage from '@/page-components/Bmc/BmcDetailPage'
 
 export default function BmcDetail({ params }: { params: { id: string } }) {
   return <BmcDetailPage />
