+++ src/app/bmc/generate/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import BmcGeneratePage from '@/pages/Bmc/BmcGeneratePage'
+import BmcGeneratePage from '@/page-components/Bmc/BmcGeneratePage'
 
 export default function BmcGenerate() {
   return <BmcGeneratePage />
