+++ src/app/chat/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import ChatPage from '@/pages/Chat'
+import ChatPage from '@/page-components/Chat'
 
 export default function ChatPageWrapper() {
   return <ChatPage />
