+++ src/app/competitor/analysis/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import CompetitorAnalysis from '@/pages/CompetitorAnalysis'
+import CompetitorAnalysis from '@/page-components/CompetitorAnalysis'
 
 export default function CompetitorAnalysisPage() {
   return <CompetitorAnalysis />
