+++ src/app/competitor/bmc-selection/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import CompetitorBmcSection from '@/pages/Competitor/BmcSelection'
+import CompetitorBmcSection from '@/page-components/Competitor/BmcSelection'
 
 export default function CompetitorBmcSelectionPage() {
   return <CompetitorBmcSection />
