+++ src/app/like-list/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import LikeList from '@/pages/LikeList'
+import LikeList from '@/page-components/LikeList'
 
 export default function LikeListPage() {
   return <LikeList />
