+++ src/app/my-profile-edit/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import EditMyProfile from '@/pages/EditProfile'
+import EditMyProfile from '@/page-components/EditProfile'
 
 export default function EditMyProfilePage() {
   return <EditMyProfile />
