+++ src/app/my-profile/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import MyProfile from '@/pages/MyProfile'
+import MyProfile from '@/page-components/MyProfile'
 
 export default function MyProfilePage() {
   return <MyProfile />
