+++ src/app/notices/[type]/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import NoticeListUpPage from '@/pages/NoticeListUp'
+import NoticeListUpPage from '@/page-components/NoticeListUp'
 
 export default function NoticeListByType({ params }: { params: { type: string } }) {
   return <NoticeListUpPage />
