+++ src/app/notices/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import NoticePage from '@/pages/Notice'
+import NoticePage from '@/page-components/Notice'
 
 export default function NoticesPage() {
   return <NoticePage />
