+++ src/app/oauth/callback/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import CallbackPage from '@/pages/Callback'
+import CallbackPage from '@/page-components/Callback'
 
 export default function OAuthCallbackPage() {
   return <CallbackPage />
