+++ src/app/oauth/fail/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import OAuthFail from '@/pages/Callback/OAuthFail'
+import OAuthFail from '@/page-components/Callback/OAuthFail'
 
 export default function OAuthFailPage() {
   return <OAuthFail />
