+++ src/app/onboarding/pre-startup/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import PreOnboarding from '@/pages/Onboarding/pre'
+import PreOnboarding from '@/page-components/Onboarding/pre'
 
 export default function PreStartupPage() {
   return <PreOnboarding />
