+++ src/page-components/Bmc/BmcDetailPage/index.tsx
