+++ src/page-components/Bmc/BmcGeneratePage/index.tsx
