+++ src/page-components/Bmc/BmcPage/index.tsx
