+++ src/page-components/Competitor/BmcSelection/index.tsx
