+++ src/page-components/Competitor/List/index.tsx
