+++ src/page-components/Competitor/index.tsx
