+++ src/page-components/CompetitorAnalysis/index.tsx
