+++ src/page-components/Main/ui/MainContent/index.tsx
