+++ src/page-components/NoticeDetail/ui/DetailContent/index.tsx
