+++ src/page-components/NoticeDetail/ui/DetailSkeleton/index.tsx
