+++ src/page-components/Onboarding/index.tsx
